refactor transaction and add transparent tx
