+++ taiga_halo2/src/lib.rs
@@ -5,12 +5,15 @@ pub mod action;
 pub mod binding_signature;
 pub mod circuit;
 pub mod constant;
+pub mod error;
 pub mod merkle_tree;
 pub mod note;
 pub mod note_encryption;
 pub mod nullifier;
 pub mod proof;
+pub mod shielded_ptx;
 pub mod transaction;
+pub mod transparent_ptx;
 pub mod utils;
 pub mod value_commitment;
 pub mod vp_vk;
